+++ BruTile/Predefined/SphericalMercatorWorldSchema.cs
@@ -23,7 +23,7 @@ namespace BruTile.Predefined
             foreach (var unitsPerPixel in unitsPerPixelArray)
             {
                 var levelId = count.ToString(CultureInfo.InvariantCulture);
-                var ms = (int) Math.Pow(count, 2) / 2;
+                var ms = (long) Math.Pow(count, 2) / 2;
 
                 // Most of the parameters to the Resolutions constructor are not used for TileSchema.
                 // This is confusing and should be changed.
